backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
gunicorn==20.1.0
pandas==1.5.3
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
xlrd==2.0.1
openpyxl==3.1.0
firebirdsql==1.2.2
fdb==2.0.2
Werkzeug==2.2.2
pyppeteer==1.0.2
requests==2.31.0
statsmodels==0.14.0